+++ android/src/main/java/com/capacitorjs/community/plugins/bluetoothle/BluetoothLe.kt
@@ -1,7 +1,13 @@
 package com.capacitorjs.community.plugins.bluetoothle
 
 import android.Manifest
-import android.bluetooth.*
+import android.annotation.SuppressLint
+import android.bluetooth.BluetoothAdapter
+import android.bluetooth.BluetoothDevice
+import android.bluetooth.BluetoothGatt
+import android.bluetooth.BluetoothGattCharacteristic
+import android.bluetooth.BluetoothManager
+import android.bluetooth.BluetoothProfile
 import android.bluetooth.le.ScanFilter
 import android.bluetooth.le.ScanResult
 import android.bluetooth.le.ScanSettings
@@ -14,15 +20,24 @@ import android.location.LocationManager
 import android.net.Uri
 import android.os.Build
 import android.os.ParcelUuid
-import android.provider.Settings.*
+import android.provider.Settings.ACTION_APPLICATION_DETAILS_SETTINGS
+import android.provider.Settings.ACTION_BLUETOOTH_SETTINGS
+import android.provider.Settings.ACTION_LOCATION_SOURCE_SETTINGS
 import androidx.core.location.LocationManagerCompat
-import com.getcapacitor.*
+import com.getcapacitor.JSArray
+import com.getcapacitor.JSObject
+import com.getcapacitor.Logger
+import com.getcapacitor.PermissionState
+import com.getcapacitor.Plugin
+import com.getcapacitor.PluginCall
+import com.getcapacitor.PluginMethod
 import com.getcapacitor.annotation.CapacitorPlugin
 import com.getcapacitor.annotation.Permission
 import com.getcapacitor.annotation.PermissionCallback
-import java.util.*
+import java.util.UUID
 
 
+@SuppressLint("MissingPermission")
 @CapacitorPlugin(
     name = "BluetoothLe",
     permissions = [
@@ -83,8 +98,7 @@ class BluetoothLe : Plugin() {
 
     @PluginMethod
     fun initialize(call: PluginCall) {
-        // Build.VERSION_CODES.S = 31
-        if (Build.VERSION.SDK_INT >= 31) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
             val neverForLocation = call.getBoolean("androidNeverForLocation", false) as Boolean
             aliases = if (neverForLocation) {
                 arrayOf(
@@ -373,7 +387,7 @@ class BluetoothLe : Plugin() {
     @PluginMethod
     fun getDevices(call: PluginCall) {
         assertBluetoothAdapter(call) ?: return
-        val deviceIds = call.getArray("deviceIds").toList<String>()
+        val deviceIds = (call.getArray("deviceIds", JSArray()) as JSArray).toList<String>()
         val bleDevices = JSArray()
         deviceIds.forEach { deviceId ->
             val bleDevice = JSObject()
@@ -426,7 +440,8 @@ class BluetoothLe : Plugin() {
     @PluginMethod
     fun createBond(call: PluginCall) {
         val device = getOrCreateDevice(call) ?: return
-        device.createBond { response ->
+        val timeout = call.getFloat("timeout", DEFAULT_TIMEOUT)!!.toLong()
+        device.createBond(timeout) { response ->
             run {
                 if (response.success) {
                     call.resolve()